refactor support bundles to only export files that would be relevant on Linux
