+++ pkg/support/config.go
@@ -4,10 +4,6 @@
 package support
 
 import (
-	"path"
-	"path/filepath"
-	"runtime"
-
 	fpath "github.com/runfinch/finch/pkg/path"
 )
 
@@ -29,24 +25,5 @@ func NewBundleConfig(finch fpath.Finch, homeDir string) BundleConfig {
 	return &bundleConfig{
 		finch:   finch,
 		homeDir: homeDir,
-	}
-}
-
-func (bc *bundleConfig) LogFiles() []string {
-	files := []string{
-		filepath.Join(bc.finch.LimaInstancePath(), "ha.stderr.log"),
-		filepath.Join(bc.finch.LimaInstancePath(), "ha.stdout.log"),
-	}
-
-	if runtime.GOOS != "windows" {
-		files = append(files, filepath.Join(bc.finch.LimaInstancePath(), "serial.log"))
-	}
-	return files
-}
-
-func (bc *bundleConfig) ConfigFiles() []string {
-	return []string{
-		path.Join(bc.finch.LimaInstancePath(), "lima.yaml"),
-		bc.finch.ConfigFilePath(bc.homeDir),
 	}
 }
